backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Cloud Platform
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-dataflow-client==0.8.6
google-cloud-resource-manager==1.12.0
google-auth==2.27.0

# Apache Beam for Dataflow pipelines
apache-beam[gcp]==2.53.0

# Data validation
pydantic==2.5.3

# Utilities
python-dateutil==2.8.2
